+++ tests/css.rs
@@ -3,7 +3,6 @@
 use svgdom::{
     Document,
     WriteOptions,
-    WriteBuffer,
 };
 
 fn write_options() -> WriteOptions {
@@ -17,7 +16,7 @@ macro_rules! test_resave {
         #[test]
         fn $name() {
             let doc = Document::from_str($in_text).unwrap();
-            assert_eq!(doc.with_write_opt(&write_options()).to_string(), $out_text);
+            assert_eq!(doc.to_string_with_opt(&write_options()), $out_text);
         }
     )
 }